Basic argparse features

+++ main.py
@@ -4,18 +4,25 @@ import shutil
 import pathlib
 import argparse
 
+# set up parser
 parser = argparse.ArgumentParser(prog="Backup Manager",
                                  description="Manages backup of everything contained in a specified directory")
-parser.add_argument("path",
-                    help="")
+parser.add_argument("-i", "--input", required=True,
+                    help="The original directory path.")
+parser.add_argument("-o", "--output", required=True,
+                    help="The backup directory path.")
 
 args = parser.parse_args()
 
-original_tree = ""
-duplicate_tree = ""
+original_tree = args.input
+duplicate_tree = args.output
 
-original_path = pathlib.PurePath(original_tree)
-duplicate_path = pathlib.PurePath(duplicate_tree)
+original_path = pathlib.Path(original_tree)
+duplicate_path = pathlib.Path(duplicate_tree)
+
+# check if provided path's exist
+original_exists = original_path.exists()
+duplicate_exists = duplicate_path.exists()
 
 copied_amount = 0
 
@@ -47,12 +54,20 @@ def ignore(visiting, contents):
     print(f"Copied: {files} Total: {copied_amount}")
     return exclude
 
-start = perf_counter()
+# perform backup or output error messages depending on whether provided path's exist
+if original_exists and duplicate_exists:
+    start = perf_counter()
 
-shutil.copytree(original_path, duplicate_path, ignore=ignore, dirs_exist_ok=True)
-# make feeback for if something exists in backup but not in original. Implement verbosity levels.
-# test with over 260 character paths copying to external drive (maybe create a thing to detect if it's necissary for those ones and notify to transfer by self)
+    shutil.copytree(original_path, duplicate_path, ignore=ignore, dirs_exist_ok=True)
+    # make feeback for if something exists in backup but not in original. Implement verbosity levels.
+    # test with over 260 character paths copying to external drive (maybe create a thing to detect if it's necissary for those ones and notify to transfer by self)
 
-end = perf_counter()
+    end = perf_counter()
 
-print(end-start)    # output time taken for backup
+    print(end-start)    # output time taken for backup
+elif not original_exists and duplicate_exists:
+    print(f"Input path {original_tree} doesn't exist.")
+elif not duplicate_exists and original_exists:
+    print(f"Output path {duplicate_path} doesn't exist.")
+else:
+    print(f"Input path {original_tree} and output path {duplicate_path} don't exist.")
